Add a 2px padding between textures in atlas.
This mitigates sampling issues when the water reflection is being drawn.
Occasionally, a slither of the texture next to an object reflected in
water could be seen at the peaks of the wave amplitude, because the
refraction shader would sample the texture next to it in the atlas.

Adding a padding in the atlas is not an ideal solution to the problem,
as it can still happen at more extreme amplitudes, and it wastes a lot
of space in the atlas. However, solving the problem in the shader may
be more complex than originally anticipated.

In TileRendererReflective, we have no knowledge of what's being drawn
before drawing it, so it's not like we can send the dimensions of the
TextureRegion as a uniform to the shader, and clamp using those. We
could simply assume that the regions are 16x16, and avoid sampling
outside of that, however that could cause issues when using textures of
different sizes, which doesn't happen right now, but might in the
future.

It might be possible to just use the UVs provided when drawing regions
from the SpriteBatch. I don't know enough about LibGDX's shaders or GLSL
to implement this myself right now, but this definitely needs to be
reviewed in the future.

I also made the Atlas Viewer window a little bigger again, so an entire
page is visible at once.